Refactor the way we inject guest author in the query
This is a major refactoring where we are moving from the approach of using the author nicename (term slug) to identify the guest author on WordPress native queries and author functions. We are moving from the slug to the term ID as negative integer. This way we keep compatibility with the sanitization methods applied in the WordPress' functions where the author ID is usually converted to an integer. That was breaking the slug to a 0 integer, loosing the data. Now we use negative integers (author term id) so we can catch that value after the sanitization and properly filter the data to match the guest author's data and make the plugin more compatible with 3rd party themes and plugins out of the box.

Fix #104, and add better compatibility for guest authors to the WordPress' functions:

* feed_links_extra
* get_author_feed_link
* get_author_posts_url
* get_the_author_meta