Nested Pages plugin: PHP Warnings
Each page on the Nested Pages screen yields a PHP Warning:

Warning: Invalid argument supplied for foreach() in F:\www\wp56\wp-content\plugins\wp-nested-pages\app\Entities\Post\PostRepository.php on line 71

Inspection of the Nested Pages code indicates that the $post->post_author property is expected to be an array of integers, but is actually an array of objects:

$terms = $post->$taxname; foreach ( $terms as $term ){ $out .= 'inf-' . $taxonomy->name . '-nps-' . $term . ' '; }

Don't register the post_author taxonomy on the Nested Pages screen. Hide the Author display/selector in the NP modal until we develop an integration.