Add get_post_authors and get_archive_author replacing get_multiple_authors Fix issue #593 where secondary authors were dropped from the first post of an author displayed on its author page. After visiting the author page the post author was reset to only the first author.
The function get_multiple_authors is deprecated now, and was replaced by two new functions:

get_post_authors, which returns the list of authors of a post (similar behavior when $archive was false).
get_archive_author, which returns the author of the current author page if on an author archive page. If not, returns false. It returns similar result as the $archive argument was true.